feat(api): Swagger용 인터페이스를 추가해서 메인 로직에서 Swagger 어노테이션을 분리한다.
